Allow multiple facets to have the same filter key
This lets multiple facets on a finder manipulate the same filter.

For example, in this case we want the Brexit checkbox, and the
top-level taxons dropdown select, to set the filter field
part_of_taxonomy_tree.

This also lets us set filter values and keys that we display
to the user be different to what we send to rummager. In
this case we display related_to_brexit=true in the URL but
send part_of_taxonomy_tree=<brexit_content_id> to rummager.

I also refactored the filter code to make it clearer (I hope).
