+++ app/lib/filter_query_builder.rb
@@ -8,7 +8,7 @@ class FilterQueryBuilder
 
   def call
     filters.select(&:active?).reduce({}) { |query, filter|
-      query.merge(filter.key => filter.value)
+      query.merge(filter.key => filter_value(query, filter))
     }
   end
 
@@ -22,12 +22,12 @@ private
 
   def build_filter(facet)
     filter_class = {
-      'checkbox' => CheckboxFilter,
-      'date' => DateFilter,
-      'hidden' => HiddenFilter,
-      'text' => TextFilter,
-      'topical' => TopicalFilter,
-      'taxon' => TaxonFilter,
+      'checkbox' => Filters::CheckboxFilter,
+      'date' => Filters::DateFilter,
+      'hidden' => Filters::HiddenFilter,
+      'text' => Filters::TextFilter,
+      'topical' => Filters::TopicalFilter,
+      'taxon' => Filters::TaxonFilter,
     }.fetch(facet['type'])
 
     params = user_params.fetch(facet['key'], nil)
@@ -35,117 +35,11 @@ private
     filter_class.new(facet, params)
   end
 
-  class Filter
-    def initialize(facet, params)
-      @facet = facet
-      @params = params
-    end
+  def filter_value(query, filter)
+    # If the same filter key is provided multiple times, provide an array
+    # of all values for that filter key
+    return Array(query[filter.key]) + Array(filter.value) if query[filter.key]
 
-    def key
-      facet['filter_key'] || facet['key']
-    end
-
-    def active?
-      value.present?
-    end
-
-    def value
-      raise NotImplementedError
-    end
-
-  private
-
-    attr_reader :facet, :params
-  end
-
-  class DateFilter < Filter
-    def value
-      serialized_values.join(",")
-    end
-
-  private
-
-    def serialized_values
-      present_values.map { |key, date|
-        "#{key}:#{date.iso8601}"
-      }
-    end
-
-    def present_values
-      parsed_values.select { |_, date|
-        date.present?
-      }
-    end
-
-    def parsed_values
-      user_values.reduce({}) { |values, (key, date_string)|
-        values.merge(key => DateParser.parse(date_string))
-      }
-    end
-
-    def user_values
-      if params.is_a?(Hash)
-        params
-      else
-        {}
-      end
-    end
-  end
-
-  class HiddenFilter < Filter
-    def value
-      params
-    end
-  end
-
-  class TextFilter < Filter
-    def value
-      Array(params)
-    end
-  end
-
-  class CheckboxFilter < Filter
-    def value
-      params
-    end
-  end
-
-  class TaxonFilter < TextFilter
-  end
-
-  class TopicalFilter < Filter
-    def value
-      return nil if params.blank?
-
-      user_has_selected_open = params.include?(facet['open_value']['value'])
-      user_has_selected_closed = params.include?(facet['closed_value']['value'])
-
-      # with both or neither selected, the filter is not used
-      if user_has_selected_open && !user_has_selected_closed
-        open_value
-      elsif user_has_selected_closed && !user_has_selected_open
-        closed_value
-      end
-    end
-
-  private
-
-    # A thing is open when it ends on a future day
-    def open_value
-      "from:#{later_than_midnight_today}"
-    end
-
-    # A thing becomes closed when it ends today or before
-    def closed_value
-      "to:#{midnight_today}"
-    end
-
-    def midnight_today
-      Time.zone.now.beginning_of_day.utc
-    end
-
-    def later_than_midnight_today
-      Time.zone.now.beginning_of_day.change(sec: 1).utc
-    end
+    filter.value
   end
 end
