+++ app/lib/filters/hidden_filter.rb
@@ -1,0 +1,7 @@
+module Filters
+  class HiddenFilter < Filter
+    def value
+      params
+    end
+  end
+end
