+++ app/lib/filters/taxon_filter.rb
@@ -1,0 +1,4 @@
+module Filters
+  class TaxonFilter < TextFilter
+  end
+end
