+++ app/lib/filters/text_filter.rb
@@ -1,0 +1,7 @@
+module Filters
+  class TextFilter < Filter
+    def value
+      Array(params)
+    end
+  end
+end
