+++ app/views/finders/_checkbox_facet.html.erb
@@ -1,16 +1,16 @@
 <div class="govuk-checkboxes">
-  <% checkbox_facet.checkboxes.each do |checkbox| %>
-    <%= render "govuk_publishing_components/components/checkbox", {
+  <div class="govuk-checkboxes__item">
+    <%# <%= render "govuk_publishing_components/components/checkbox", {
       name: checkbox_facet.name,
-      id: checkbox.id,
+      id: checkbox_facet.id,
       controls: "js-search-results-info",
       margin_bottom: true,
       item: {
-        label: checkbox.checkbox_label,
-        value: checkbox.value,
-        checked: checkbox_facet.checked?(checkbox),
-        data_attributes: checkbox.data
+        label: checkbox_facet.checkbox_label,
+        value: checkbox_facet.value,
+        checked: checkbox_facet.is_checked?,
+        data_attributes: checkbox_facet.data
       }
-    } %>
-  <% end %>
+    } %> %>
+  </div>
 </div>
