+++ spec/models/checkbox_facet_spec.rb
@@ -1,37 +1,15 @@
 require "spec_helper"
 
 describe CheckboxFacet do
-  let(:checkboxes) {
-    [
-      {
-        'label' => "Allowed value 1",
-        'checkbox_label' => "Show allowed value 1",
-        'value' => "allowed-value-1"
-      },
-      {
-        'label' => "Allowed value 2",
-        'checkbox_label' => "Show allowed value 2",
-        'value' => "allowed-value-2"
-      },
-      {
-        'label' => "Remittals",
-        'checkbox_label' => "Show allowed value 3",
-        'value' => "remittals"
-      }
-    ]
-  }
-
   let(:facet_data) {
     {
       'type' => "checkbox",
-      'key' => "test_values",
+      'key' => "show_extra_information",
+      'label' => "Show extra information",
+      'checkbox_label' => "Show extra information",
+      'value' => "yes",
       'preposition' => "of value",
-      'checkboxes' => checkboxes,
     }
-  }
-
-  let(:checkbox) {
-    Checkbox.new(checkboxes.first)
   }
 
   subject { CheckboxFacet.new(facet_data) }
@@ -42,31 +20,30 @@ describe CheckboxFacet do
     end
 
     context "single value" do
-      let(:value) { ["allowed-value-1"] }
+      let(:value) { "yes" }
 
       specify {
         expect(subject.sentence_fragment['preposition']).to eql("of value")
-        expect(subject.sentence_fragment['values'].first['label']).to eql("Allowed value 1")
-        expect(subject.sentence_fragment['values'].first['parameter_key']).to eql("test_values")
+        expect(subject.sentence_fragment['values'].first['label']).to eql("Show extra information")
+        expect(subject.sentence_fragment['values'].first['parameter_key']).to eql("show_extra_information")
       }
     end
 
-    context "multiple values" do
-      let(:value) { ["allowed-value-1", "allowed-value-2"] }
+    context "when multiple values are provided" do
+      context "when a value is provided" do
+        let(:value) { true }
 
-      specify {
-        expect(subject.sentence_fragment['preposition']).to eql("of value")
-        expect(subject.sentence_fragment['values'].first['label']).to eql("Allowed value 1")
-        expect(subject.sentence_fragment['values'].first['parameter_key']).to eql("test_values")
+        specify {
+          expect(subject.sentence_fragment['preposition']).to eql("of value")
+          expect(subject.sentence_fragment['values'].count).to eql 1
+          expect(subject.sentence_fragment['values'].first['parameter_key']).to eql("show_extra_information")
+        }
+      end
 
-        expect(subject.sentence_fragment['values'].last['label']).to eql("Allowed value 2")
-        expect(subject.sentence_fragment['values'].last['parameter_key']).to eql("test_values")
-      }
-    end
-
-    context "disallowed values" do
-      let(:value) { ["disallowed-value-1, disallowed-value-2"] }
-      specify { expect(subject.sentence_fragment).to be_nil }
+      context "when no value is provided" do
+        let(:value) { nil }
+        specify { expect(subject.sentence_fragment).to be_nil }
+      end
     end
   end
 
@@ -76,23 +53,16 @@ describe CheckboxFacet do
     end
 
     context "checkbox is selected" do
-      let(:value) { %w(allowed-value-1) }
+      let(:value) { "yes" }
       specify {
-        expect(subject.checked?(checkbox)).to eql(true)
+        expect(subject.is_checked?).to eql(true)
       }
     end
 
     context "checkbox is not selected" do
-      let(:value) { [] }
+      let(:value) { nil }
       specify {
-        expect(subject.checked?(checkbox)).to eql(false)
-      }
-    end
-
-    context "another checkbox is selected" do
-      let(:value) { %w(allowed-value-2) }
-      specify {
-        expect(subject.checked?(checkbox)).to eql(false)
+        expect(subject.is_checked?).to eql(false)
       }
     end
   end
